Created a new file, match_predictor.py, to allow users to input pre-match details (Date, Teams, and Best Of) and receive predictions for scores, outcomes, and maps. This tool was designed primarily for debugging, enabling the testing of individual predictions instead of relying on large .csv datasets.
Initial tests revealed areas for improvement in the model. Notable issues include predictions of the same map being played multiple times in a Bo3/Bo5, which is invalid under VCT rules, and unrealistic score predictions like "1.48 - 0.91" for a Bo3. These results highlight the need for further refinements to the model to better align predictions with real-world constraints and match dynamics.